Move alert processing into a separate service
