+++ server/src/services/incidents/incidents.service.ts
@@ -73,7 +73,7 @@ export default function (app: Application): void {
   };
 
   // Initialize our service with any options it requires
-  app.use('/incidents', new Incidents(options));
+  app.use('/incidents', new Incidents(options, app));
 
   // Get our initialized service so that we can register hooks
   const service = app.service('incidents');
